move lane_conforming to another file and update code according to suggested comments
