+++ rs/utils/src/distance/l2.rs
@@ -1,4 +1,3 @@
-use std::marker::PhantomData;
 use std::ops::Mul;
 use std::simd::num::SimdFloat;
 use std::simd::{f32x16, f32x4, f32x8, LaneCount, Simd, SupportedLaneCount};
@@ -87,27 +86,10 @@ impl DistanceCalculator for L2DistanceCalculator {
                 *acc += diff.mul(diff);
             });
     }
-}
-
-/// Calculator where we know in advance that the dimension of vectors is a multiple of LANES.
-/// This skips a bunch of checks and allows for a more efficient implementation.
-pub struct LaneConformingDistanceCalculator<const LANES: usize, D: DistanceCalculator + Send + Sync>
-where
-    LaneCount<LANES>: SupportedLaneCount,
-{
-    _marker: PhantomData<D>,
-}
-
-impl<const LANES: usize, D: DistanceCalculator + Send + Sync> CalculateSquared
-    for LaneConformingDistanceCalculator<LANES, D>
-where
-    LaneCount<LANES>: SupportedLaneCount,
-{
+    
     #[inline(always)]
-    fn calculate_squared(a: &[f32], b: &[f32]) -> f32 {
-        let mut simd = Simd::<f32, LANES>::splat(0.0);
-        D::accumulate_lanes(a, b, &mut simd);
-        simd.reduce_sum()
+    fn outermost_op(x: f32) -> f32 {
+        x
     }
 }
