+++ rs/utils/src/distance/mod.rs
@@ -1,2 +1,3 @@
 pub mod dot_product;
 pub mod l2;
+pub mod lane_conforming;
